+++ tests/components/ecobee/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Ecobee integration."""
