+++ tests/components/google/test_tts.py
@@ -12,7 +12,7 @@ from homeassistant.setup import setup_component
 from tests.common import (
     get_test_home_assistant, assert_setup_component, mock_service)
 
-from .test_init import mutagen_mock  # noqa
+from tests.components.tts.test_init import mutagen_mock  # noqa
 
 
 class TestTTSGooglePlatform:
