+++ tests/components/mochad/test_switch.py
@@ -6,7 +6,7 @@ import pytest
 
 from homeassistant.setup import setup_component
 from homeassistant.components import switch
-from homeassistant.components.switch import mochad
+from homeassistant.components.mochad import switch as mochad
 
 from tests.common import get_test_home_assistant
