+++ tests/components/webostv/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the WebOS TV integration."""
