+++ tests/components/xiaomi_miio/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Xiaomi Miio integration."""
